User credits increasing
