+++ app/config/stripe.ts
@@ -1,0 +1,4 @@
+// At the top of your file, define the credit mapping
+export const PRICE_TO_CREDITS: Record<string, number> = {
+  price_1QHVBkFMpb9ac8oli1eIxXb1: 1, // Replace with your actual Stripe price IDs
+};
